// source/frontend/src/models/api/generated/ApiGen_Concepts_ProjectProduct.ts
/**
 * File autogenerated by TsGenerator.
 * Do not manually modify, changes made to this file will be lost when this file is regenerated.
 */
import { ApiGen_Base_BaseAudit } from './ApiGen_Base_BaseAudit';
import { ApiGen_Concepts_Product } from './ApiGen_Concepts_Product';
import { ApiGen_Concepts_Project } from './ApiGen_Concepts_Project';

// LINK: @backend/apimodels/Models/Concepts/Project/ProjectProductModel.cs
export interface ApiGen_Concepts_ProjectProduct extends ApiGen_Base_BaseAudit {
  id: number;
  projectId: number;
  product: ApiGen_Concepts_Product | null;
  productId: number;
  project: ApiGen_Concepts_Project | null;
}